+++ types/inputsOutputs.d.ts
@@ -1,5 +1,5 @@
 /*
- * Input-Output that transit between the builder and the notifier, as the PDF variable.
+ * Input-Output that transits between the builder and the notifier, as the PDF variable.
  * Used as input, then set to "" by the output. See the bpmn for details
  */
 interface _PDFBuilderToNotifInputOutput {