+++ types/variables.d.ts
@@ -8,9 +8,9 @@ import {TDateISO, TFlatDate} from "./dates";
 
 // these are created at the start of the all instances
 export interface PhDAssessStartEventVariables {
-  dashboardDefinition: StepsDefinition | undefined;  // defines the rules/flow of the dashboard
-  // encrypted stack of NotificationLog entries. As it is encrypted, it is
-  // a string. Once decrypted and JSON.parsed, it becomes
+  dashboardDefinition: StepsDefinition | undefined;  // defines the dashboard rules/flow
+  // encrypted stack of NotificationLog entries; as it is encrypted, it is
+  // a string; once decrypted and JSON.parsed, it becomes
   // a type of NotificationLog[]
   notificationLogs: string[] | undefined;
 }